+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Changed
 - Move hyperapp from peerDependencies to dependencies.
 - Remove `<apollo-query/>` element from DOM tree.
+- `variables` render prop of specifies the variable object the query was called with.
 
 ## [0.2.0] - 2018-03-29
 ### Changed